+++ src/library/DIPS.Mobile.UI/Components/ListItems/Options/Title/TitleOptions.cs
@@ -9,6 +9,7 @@ public partial class TitleOptions : ListItemOptions
         if(listItem.TitleLabel is null)
             return;
 
+        listItem.TitleLabel.SetBinding(Label.TextProperty, new Binding(nameof(ListItem.Title), source: listItem));
         listItem.TitleLabel.SetBinding(Label.FontAttributesProperty, new Binding(nameof(FontAttributes), source: this));
         listItem.TitleLabel.SetBinding(VisualElement.StyleProperty, new Binding(nameof(Style), source: this));
         listItem.TitleLabel.SetBinding(Label.TextColorProperty, new Binding(nameof(TextColor), source: this));
@@ -17,7 +18,7 @@ public partial class TitleOptions : ListItemOptions
         listItem.TitleLabel.SetBinding(Label.LineBreakModeProperty, new Binding(nameof(LineBreakMode), source: this));
         listItem.TitleLabel.SetBinding(View.MarginProperty, new Binding(nameof(Margin), source: this));
         listItem.TitleLabel.SetBinding(Label.FormattedTextProperty, new Binding(nameof(FormattedText), source: this));
-        listItem.TitleLabel.SetBinding(Label.MaximumWidthRequestProperty, new Binding(nameof(MaxWidth), source: this));
+        listItem.TitleLabel.SetBinding(VisualElement.MaximumWidthRequestProperty, new Binding(nameof(MaxWidth), source: this));
         
         if (MaxLines > -1) //We can not trigger property changed for this if its -1 because it causes bugs on Android.
         {